[refactor] Carousel다이어트
- CarouselViewConstants 생성하여 enum, 상수값 옮김
